requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.1.8
constantly==23.10.4
cryptography==45.0.6
daily-python==0.19.4
daphne==4.2.1
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
edge-tts==7.2.0
exceptiongroup==1.3.0
frozenlist==1.7.0
fuzzywuzzy==0.18.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
msgpack==1.1.1
multidict==6.6.3
openai==0.28.0
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
redis==6.3.0
requests==2.32.4
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SpeechRecognition==3.14.3
sqlparse==0.5.3
stripe==12.3.0
tabulate==0.9.0
tomli==2.2.1
tqdm==4.67.1
twilio==9.6.5
Twisted==25.5.0
txaio==23.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vapi-python==0.1.9
yarl==1.20.1
zope.interface==7.2
